# Optimize Message handler registry hot path