@SimonCropp 2016-06-27
what is the total memory when we have a large number of types?