@danielmarbach 2016-06-28
@SimonCropp I will also add an overall size comparison. But in throughput mode the modifications already rock and I haven't eliminated the `IsAssignableFrom` call yet.