@danielmarbach 2016-06-28
Test used https://github.com/danielmarbach/MicroBenchmarks/blob/master/MicroBenchmarks/NServiceBus/MessageHandlerRegistry.cs