@danielmarbach 2016-06-29
@SimonCropp I haven't fixed the test yet but I updated the size data. The test I used was this one 

https://github.com/danielmarbach/MicroBenchmarks/blob/master/MicroBenchmarks/NServiceBus/MessageHandlerRegistry.cs#L112

Basically creating the registry, registering handlers and do one resolve for a given message type. As you can see the new version in that comparison uses slightly more memory (due to Type instead of RuntimeTypeHandle usage) but overall the perf and memory traffic is much better in the optimized version.