@danielmarbach 2016-06-29
@Particular/nservicebus-maintainers this is ready for review