@bording 2016-06-30
@danielmarbach I'm assuming the "struct vs class" in your test results are referring to `DelegateHolder`?

One potential concern here would be around the fact that this is now a mutable struct that contains members that are reference types. Given that this is a private class, the dangers of this are constrained, but Weird Things can happen with mutable structs. Have you verified that everything is ok in this scenario?