@danielmarbach 2016-06-30
Yes. When handler are registered the holder is created and cache forever
and only ever read from

On Thursday, June 30, 2016, Brandon Ording <notifications@github.com> wrote:

> @danielmarbach <https://github.com/danielmarbach> I'm assuming the
> "struct vs class" in your test results are referring to DelegateHolder?
>
> One potential concern here would be around the fact that this is now a
> mutable struct that contains members that are reference types. Given that
> this is a private class, the dangers of this are constrained, but Weird
> Things can happen with mutable structs. Have you verified that everything
> is ok in this scenario?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/Particular/NServiceBus/pull/3878#issuecomment-229752644>,
> or mute the thread
> <https://github.com/notifications/unsubscribe/AAKosqgkz8RlfevdvtL7wMSkx-yeI823ks5qRA8vgaJpZM4I-LTG>
> .
>