@timbussmann 2016-07-06
Looks like all comments were addressed, @danielmarbach this is good to merge, right?